ARM: tegra: register pinmux and gpio devices before pinmux init
In the original commit, it looks like some of this commit was
reverted so I've only added the line to the board-grouper.c file.
Indeed it looks like this is fine as that line is still in
board-kai.c. This should mean everything will have sorted itself
out once the rest is updated